tests: drop compare test
This was mostly useless, originally had the idea to structure a lot of
tests around this but found it to not be worth it
