+++ crates/agentgateway/tests/common/mod.rs
@@ -1,4 +1,3 @@
-pub mod compare;
 pub mod gateway;
 pub mod hbone_server;
 pub mod mock_ca_server;
