+++ crates/agentgateway/tests/tests/mod.rs
@@ -1,4 +1,3 @@
-mod envoycompare;
 mod hbone;
 pub mod llm;
 mod smoke;
